requirements.txt
# Core backend
fastapi
uvicorn
pydantic
python-dotenv
requests
python-multipart

# UI
streamlit

# Parsing
pypdf
python-docx
beautifulsoup4
lxml   # optional but recommended


# Lightweight numeric ops (embeddings + similarity)
numpy

# Optional: Postgres + pgvector (comment out if not using DB now)
psycopg[binary]>=3.2
